Complete refactoring to support async lab creation and deletion, and adding support for multiple collections of domains to assign (via Resource Groups)

+++ LabManageJob/Functions.cs
@@ -6,6 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Lab.Common;
 using Lab.Common.Repo;
+using Lab.Data.Helper;
 using Lab.Data.Models;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Host;
@@ -26,28 +27,37 @@ namespace LabManageJob
         /// <param name="log"></param>
         public static async Task ProcessQueueMessage([QueueTrigger("%logqueue%")] string message, TextWriter log)
         {
+            var data = JsonConvert.DeserializeObject<LabQueueDTO>(message);
+            var lab = await LabRepo.GetLab(data.LabId);
+            if (lab.State == LabState.Error)
+                return;
+
             try
             {
-                var data = JsonConvert.DeserializeObject<LabQueueDTO>(message);
-                var lab = await LabRepo.GetLab(data.LabId);
-                switch (data.Operation)
+                switch (lab.State)
                 {
-                    case "RemoveLab":
+                    case LabState.QueuedToDelete:
+                        lab.State = LabState.Deleting;
+                        await LabRepo.UpdateLab(lab, data.UserName);
                         await LabRepo.RemoveLabAssignments(lab);
                         break;
-                    case "AddLab":
+                    case LabState.Queued:
+                        lab.State = LabState.Creating;
+                        await LabRepo.UpdateLab(lab, data.UserName);
                         await LabRepo.AddLabAssignments(lab);
                         break;
                     default:
                         return;
                 }
-                await Logging.WriteMessageToErrorLog(string.Format("{4} called \"{5}\" for {0} zones associated with lab \"{1} ({2})\", using RG \"{3}\"", lab.AttendeeCount, lab.City, lab.LabDate.ToShortDateString(), lab.DnsZoneRG, data.UserName, data.Operation));
+                await Logging.WriteMessageToErrorLog(string.Format("{4} called \"{5}\" for {0} zones associated with lab \"{1} ({2})\", using RG \"{3}\"", lab.AttendeeCount, lab.City, lab.LabDate.ToShortDateString(), lab.DnsZoneRG, data.UserName, lab.State.ToString()));
                 log.WriteLine(message);
             }
             catch (Exception ex)
             {
                 await Logging.WriteDebugInfoToErrorLog("Error processing queue", ex);
                 log.WriteLine(Logging.GetExceptionMessageString(ex));
+                lab.State = LabState.Error;
+                await LabRepo.UpdateLab(lab, data.UserName);
                 throw ex;
             }
         }
